--- data-models/User.js
@@ -1,41 +1,0 @@
-import mongoose from "mongoose";
-
-const UserSchema = new mongoose.Schema(
-    {
-        firstName: {
-            type: String,
-            require: true,
-            min: 2,
-            max: 50,
-        },
-        lastName: {
-            type: String,
-            require: true,
-            min: 2,
-            max: 50,
-        },
-        email: {
-            type: String,
-            require: true,
-            max: 50,
-            unique: true,
-        },
-        password: {
-            type: String,
-            require: true,
-            min: 5,
-        },
-        picturePath: {
-            type: String,
-            default: []
-        },
-        location: String,  
-        occupation: String,
-        viewedProfile: Number,
-        impressions: Number,
-    },
-    { timestamps: true }
-);
-
-const User = mongoose.model("User", UserSchema)
-export default User;
